Use test_name as container name and fix test to use new distro API

+++ pyci/tests/shell/subcommands/test_pack.py
@@ -188,25 +188,29 @@ def test_wheel_not_python_project(pack, binary):
     'run:Ubuntu:16.04',
     'run:Ubuntu:14.04'
 ])
-def test_binary_cross_distribution_wheel(repo_version, repo_path, build_distro_id, run_distro_id):
+def test_binary_cross_distribution_wheel(repo_version, repo_path, test_name, build_distro_id, run_distro_id):
 
-    build_distro = distros.from_string(build_distro_id)
-    run_distro = distros.from_string(run_distro_id)
+    build_distro = distros.from_string(test_name, build_distro_id)
+    run_distro = distros.from_string(test_name, run_distro_id)
 
-    # cant build packages on a distro with no python installed
-    assert build_distro.has_python
-
-    if run_distro.has_python:
+    if run_distro.python_version is not None:
         # wheels can only be built on distros with python installed
-        expected_result = 'py_ci-{}-{}-none-any.whl'.format(
+        expected_output = 'py_ci-{}-{}-none-any.whl'.format(
             repo_version,
             'py2' if run_distro.python_version.startswith('2') else 'py3')
     else:
-        expected_result = 'Python installation not found in PATH'
+        expected_output = 'Python installation not found in PATH'
 
-    local_binary_path = build_distro.binary(repo_path)
+    local_binary_path = None
 
     try:
+
+        build_distro.boot()
+
+        local_binary_path = build_distro.binary(repo_path)
+
+        run_distro.boot()
+
         remote_binary_path = run_distro.add(local_binary_path)
         remote_repo_path = run_distro.add(repo_path)
 
@@ -217,10 +221,15 @@ def test_binary_cross_distribution_wheel(repo_version, repo_path, build_distro_i
             # http://click.palletsprojects.com/en/6.x/python3/
             locale_setup = 'export LC_ALL=C.UTF-8 && export LANG=C.UTF-8 &&'
 
-        result = run_distro.run('chmod +x {0} && {2} {0} pack --path {1} wheel'
-                                .format(remote_binary_path, remote_repo_path, locale_setup), exit_on_failure=False)
+        result = run_distro.run('chmod +x {} && {} {} pack --path {} wheel'
+                                .format(remote_binary_path, locale_setup, remote_binary_path, remote_repo_path),
+                                exit_on_failure=False)
 
-        assert expected_result in result.std_out
+        assert expected_output in result.std_out
 
     finally:
-        os.remove(local_binary_path)
+        if local_binary_path:
+            os.remove(local_binary_path)
+        build_distro.shutdown()
+        run_distro.shutdown()
+
